refactor: improve loudness range calculations by linear interpolation
